Incorrect import of math module. In Python3 .__doc__ attribute lives on the module itself and not on __func__.__doc__. UserDict.DictMixin is collections.MutableMapping in Python3. Plain injection of member function, random_selection(), on ext.mersenne_twister class without using boost.python.injector. long doesn't exist in Python3.

+++ scitbx/random/__init__.py
@@ -9,7 +9,7 @@ boost.python.import_ext("scitbx_random_ext")
 import scitbx_random_ext
 
 builtin_int = __builtins__["int"]
-builtin_long = __builtins__["long"]
+builtin_long = __builtins__["int"] # long doesn't exist in Python 3
 
 def get_random_seed():
   try:
